feat: rename library, prepare for pypi
