+++ grabit.py
@@ -2,10 +2,10 @@
 # requires-python = ">=3.11"
 # dependencies = [
 #   "click>=8.1.0,<8.2",
-#   "grabit_lib>=1.0.0",
+#   "grabit-md>=1.0.0",
 # ]
 # ///
-from grabit_lib.cli import main
+from grabit_md.cli import main
 
 if __name__ == "__main__":
     main()
