+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "grabit_lib"
+name = "grabit-md"
 version = "1.0.0"
 description = "Grabit is a library that allows you to download web pages, extract their readable content, convert it to Markdown, and save it locally."
 readme = "README.md"
@@ -17,7 +17,7 @@ classifiers = [
 dependencies = ["readabilipy==0.3.0", "markdownify==1.2.0", "PyYAML==6.0.2", "requests==2.32.5", "mdformat==0.7.22", "click>=8.1.0,<8.2"]
 
 [project.scripts]
-grabit = "grabit_lib.cli:main"
+grabit = "grabit_md.cli:main"
 
 [dependency-groups]
 # uv-only groups for local development; not published as pip extras
