+++ src/grabit_md/__init__.py
@@ -1,10 +1,10 @@
 from importlib.metadata import PackageNotFoundError, version
 
-from grabit_lib.core import GrabitError, OutputFormat
-from grabit_lib.grabber import Grabber
+from grabit_md.core import GrabitError, OutputFormat
+from grabit_md.grabber import Grabber
 
 try:
-    __version__ = version("grabit_lib")
+    __version__ = version("grabit-md")
 except PackageNotFoundError:
     # Fallback when running from source without installation
     __version__ = "dev"
