+++ src/grabit_md/grabbers/reddit_grabber.py
@@ -1,11 +1,11 @@
 import json
 from urllib.parse import urlparse, urlunparse
 
-from grabit_lib import GrabitError
-from grabit_lib.core import RenderFlags
-from grabit_lib.core.downloader import download_html_content
-from grabit_lib.core.output_format import OutputFormat, OutputFormatList
-from grabit_lib.grabbers.base_grabber import BaseGrabber
+from grabit_md import GrabitError
+from grabit_md.core import RenderFlags
+from grabit_md.core.downloader import download_html_content
+from grabit_md.core.output_format import OutputFormat, OutputFormatList
+from grabit_md.grabbers.base_grabber import BaseGrabber
 
 
 class RedditGrabber(BaseGrabber):
